refactor(oauthdb): Split OAuthDB implementation across multiple files.
This lets us share validation objects between the external routes and
internal API calls, and should make things more manageable when we
come to add addional routes.
